make sure asset layers are displayed on map load
Refactor the asset layer displaying code so it doesn't wait till a
category is chosen before determining which layers to show. As part of
this we set the relevant bodies as a data attribute so they're available
without another call to the server. This also means that non category
linked assets are shown initially.

As part of this create a new VectorAsset class to handle the visibility
changes for asset layers as we need them to fire on both category update
and category extras update and this makes that all a bit tidier.

This also moved the bodies property up to the fixmystreet object from
the fixmystreet.map object because the map object may not be available
when we set this.